전처리, setting 부분도 Experiment 객체로 이동
