+++ accutuning_client/object.py
@@ -94,6 +94,58 @@ class Experiment(dict):
         self.update(result.get('runtime'))
         self._update_timestamp()
 
+    def preprocessor_config_recommend(self):
+        '''
+        데이터셋을 기반으로 Preprocessor 방법을 자동으로 추천받아 preprocessor config를 변경합니다.
+        '''
+        query = '''
+            mutation patchRecmdConfig($id:ID!) {
+                patchRecommendationConfig (id: $id) {
+                        dataset {
+                            id
+                            processingStatus
+                    }
+                }
+            }
+        '''
+        self._graphql.execute(query, {'id': self.get('dataset.id')})
+
+    def preprocess(self):
+        '''
+        지정된 preprocessor config 설정대로 전처리를 실시합니다.
+        '''
+        query = '''
+            mutation preprocess($id:ID!) {
+                preprocess (id: $id) {
+                    dataset {
+                        id
+                        processingStatus
+                    }
+                    error
+                    errorMessage
+                }
+            }
+        '''
+        self._graphql.execute(query, {'id': self.get('dataset.id')})
+
+    def set_runtime_settings(self, estimator_type, metric, target_column_name):
+        '''
+        '''
+        query = '''
+            mutation ($id: ID!, $input: PatchRuntimeInput!) {
+                patchRuntime(id: $id, input: $input) {
+                    runtime {
+                        id
+                        name
+                        estimatorType
+                        metric
+                        targetColumnName
+                    }
+                }
+            }
+        '''
+        self._graphql.execute(query, {'id': self.get('id'), 'input': {'estimatorType': estimator_type, 'metric': metric, 'targetColumnName': target_column_name}})
+
     def run(self):
         """Run AutoML
